CI Updates

See merge request tsmc40r/brainpower/accelerator!8

Former-commit-id: 353ea3373c9c8bd28a4d9e240ce4d21b7d35a52a

+++ test/common/Utils.h
@@ -9,19 +9,19 @@
 #include "src/ArchitectureParams.h"
 #include "test/common/UniversalPosit.h"
 
-int compare_arrays(INPUT_DATATYPE *matrixA, INPUT_DATATYPE *matrixB,
+float compare_arrays(INPUT_DATATYPE *matrixA, INPUT_DATATYPE *matrixB,
                    size_t size, std::string filename, bool accType);
-int compare_arrays(INPUT_DATATYPE *matrixA, float *matrixB, size_t size,
+float compare_arrays(INPUT_DATATYPE *matrixA, float *matrixB, size_t size,
                    std::string filename, bool accType);
 #ifndef NO_UNIVERSAL
-int compare_arrays(INPUT_DATATYPE *matrixA, UniversalPosit *matrixB,
+float compare_arrays(INPUT_DATATYPE *matrixA, UniversalPosit *matrixB,
                    size_t size, std::string filename, bool accType);
-int compare_arrays(UniversalPosit *matrixA, UniversalPosit *matrixB,
+float compare_arrays(UniversalPosit *matrixA, UniversalPosit *matrixB,
                    size_t size, std::string filename, bool accType);
-int compare_arrays(UniversalPosit *matrixA, float *matrixB, size_t size,
+float compare_arrays(UniversalPosit *matrixA, float *matrixB, size_t size,
                    std::string filename, bool accType);
 #endif
-int compare_arrays(float *matrixA, float *matrixB, size_t size,
+float compare_arrays(float *matrixA, float *matrixB, size_t size,
                    std::string filename, bool accType);
 
 int validateMapping(SimplifiedParams params);
